added tests cases
