+++ test/adapters/configuration.test.js
@@ -24,15 +24,12 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-const { utils } = require("ethers");
 const {
   sha3,
   toBN,
   fromAscii,
   toWei,
-  unitPrice,
   DAI_TOKEN,
-  UNITS,
   LOOT,
   ZERO_ADDRESS,
 } = require("../../utils/contract-util");
@@ -46,11 +43,8 @@ const {
   accounts,
   expectRevert,
   expect,
-  OLToken,
   web3,
 } = require("../../utils/oz-util");
-
-const { onboardingNewMember } = require("../../utils/test-util");
 
 const owner = accounts[1];
 const proposalCounter = proposalIdGenerator().generator;
@@ -513,343 +507,6 @@ describe("Adapter - Configuration", () => {
         data: fromAscii("should go to fallback func"),
       }),
       "revert"
-    );
-  });
-
-  it("should be possible to update a DAO configuration if you are a member and a maintainer that holds an external governance token", async () => {
-    const maintainer = accounts[5];
-    // Issue OpenLaw ERC20 Basic Token for tests, only DAO maintainer will hold this token
-    const tokenSupply = toBN(100000);
-    const oltContract = await OLToken.new(tokenSupply);
-
-    // Transfer OLTs to the maintainer account
-    await oltContract.transfer(maintainer, toBN(1));
-    const maintainerBalance = await oltContract.balanceOf.call(maintainer);
-    expect(maintainerBalance.toString()).equal("1");
-
-    const { dao, adapters } = await deployDefaultDao({
-      owner,
-      maintainerTokenAddress: oltContract.address,
-    });
-    const voting = adapters.voting;
-    const configuration = adapters.configuration;
-    const configKey = sha3(
-      web3.utils.encodePacked(
-        "governance.role.",
-        utils.getAddress(configuration.address)
-      )
-    );
-
-    // Make sure the governance token configuration was created
-    const governanceToken = await dao.getAddressConfiguration(configKey);
-    expect(governanceToken).equal(oltContract.address);
-
-    // Onboard the maintainer as a DAO member
-    await onboardingNewMember(
-      getProposalCounter(),
-      dao,
-      adapters.onboarding,
-      voting,
-      maintainer,
-      owner,
-      unitPrice,
-      UNITS
-    );
-
-    const key = sha3("key");
-    const proposalId = getProposalCounter();
-
-    // The maintainer submits a new configuration proposal
-    await configuration.submitProposal(
-      dao.address,
-      proposalId,
-      [
-        {
-          key: key,
-          numericValue: 99,
-          addressValue: ZERO_ADDRESS,
-          configType: 0,
-        },
-      ],
-      [],
-      { from: maintainer, gasPrice: toBN("0") }
-    );
-
-    let value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("0");
-
-    // The maintainer votes on the new proposal
-    await voting.submitVote(dao.address, proposalId, 1, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    await advanceTime(10000);
-
-    // The maintainer processes on the new proposal
-    await configuration.processProposal(dao.address, proposalId, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("99");
-  });
-
-  it("should be possible to update a DAO configuration if you are a member and a maintainer that holds an internal governance token", async () => {
-    const maintainer = accounts[5];
-    const { dao, adapters } = await deployDefaultDao({
-      owner,
-      maintainerTokenAddress: UNITS, // if the member holds any UNITS he is a maintainer
-    });
-    const voting = adapters.voting;
-    const configuration = adapters.configuration;
-    const configKey = sha3(
-      web3.utils.encodePacked(
-        "governance.role.",
-        utils.getAddress(configuration.address)
-      )
-    );
-
-    // Make sure the governance token configuration was created
-    const governanceToken = await dao.getAddressConfiguration(configKey);
-    expect(governanceToken).equal(utils.getAddress(UNITS));
-
-    // Onboard the maintainer as a DAO member
-    await onboardingNewMember(
-      getProposalCounter(),
-      dao,
-      adapters.onboarding,
-      voting,
-      maintainer,
-      owner,
-      unitPrice,
-      UNITS
-    );
-
-    const key = sha3("key");
-    const proposalId = getProposalCounter();
-
-    // The maintainer submits a new configuration proposal
-    await configuration.submitProposal(
-      dao.address,
-      proposalId,
-      [
-        {
-          key: key,
-          numericValue: 99,
-          addressValue: ZERO_ADDRESS,
-          configType: 0,
-        },
-      ],
-      [],
-      { from: maintainer, gasPrice: toBN("0") }
-    );
-
-    let value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("0");
-
-    // The maintainer votes on the new proposal
-    await voting.submitVote(dao.address, proposalId, 1, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    await advanceTime(10000);
-
-    // The maintainer processes on the new proposal
-    await configuration.processProposal(dao.address, proposalId, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("99");
-  });
-
-  it("should be possible to update a DAO configuration if you are a member and a maintainer that holds the default governance token", async () => {
-    const maintainer = accounts[5];
-    const { dao, adapters } = await deployDefaultDao({
-      owner,
-      // if the member holds any UNITS that represents the default governance token,
-      // the member is considered a maintainer.
-      defaultMemberGovernanceToken: UNITS,
-    });
-    const voting = adapters.voting;
-    const configuration = adapters.configuration;
-    const configKey = sha3(web3.utils.encodePacked("governance.role.default"));
-
-    // Make sure the governance token configuration was created
-    const governanceToken = await dao.getAddressConfiguration(configKey);
-    expect(governanceToken).equal(utils.getAddress(UNITS));
-
-    // Onboard the maintainer as a DAO member
-    await onboardingNewMember(
-      getProposalCounter(),
-      dao,
-      adapters.onboarding,
-      voting,
-      maintainer,
-      owner,
-      unitPrice,
-      UNITS
-    );
-
-    const key = sha3("key");
-    const proposalId = getProposalCounter();
-
-    // The maintainer submits a new configuration proposal
-    await configuration.submitProposal(
-      dao.address,
-      proposalId,
-      [
-        {
-          key: key,
-          numericValue: 99,
-          addressValue: ZERO_ADDRESS,
-          configType: 0,
-        },
-      ],
-      [],
-      { from: maintainer, gasPrice: toBN("0") }
-    );
-
-    let value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("0");
-
-    // The maintainer votes on the new proposal
-    await voting.submitVote(dao.address, proposalId, 1, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    await advanceTime(10000);
-
-    // The maintainer processes on the new proposal
-    await configuration.processProposal(dao.address, proposalId, {
-      from: maintainer,
-      gasPrice: toBN("0"),
-    });
-
-    value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("99");
-  });
-
-  it("should not be possible to update a DAO configuration if you are a maintainer but not a member", async () => {
-    const maintainer = accounts[5]; // not a member
-
-    // Issue OpenLaw ERC20 Basic Token for tests, only DAO maintainer will hold this token
-    const tokenSupply = toBN(100000);
-    const oltContract = await OLToken.new(tokenSupply);
-
-    // Transfer OLTs to the maintainer account
-    await oltContract.transfer(maintainer, toBN(1));
-    const maintainerBalance = await oltContract.balanceOf.call(maintainer);
-    expect(maintainerBalance.toString()).equal("1");
-
-    const { dao, adapters } = await deployDefaultDao({
-      owner,
-      maintainerTokenAddress: oltContract.address, // only holders of the OLT token are considered maintainers
-    });
-    const configuration = adapters.configuration;
-    const configKey = sha3(
-      web3.utils.encodePacked(
-        "governance.role.",
-        utils.getAddress(configuration.address)
-      )
-    );
-
-    // Make sure the governance token configuration was created
-    const governanceToken = await dao.getAddressConfiguration(configKey);
-    expect(governanceToken).equal(oltContract.address);
-
-    let key = sha3("key");
-
-    const proposalId = getProposalCounter();
-    //Submit a new configuration proposal
-    await expectRevert(
-      configuration.submitProposal(
-        dao.address,
-        proposalId,
-        [
-          {
-            key: key,
-            numericValue: 99,
-            addressValue: ZERO_ADDRESS,
-            configType: 0,
-          },
-        ],
-        [],
-        { from: maintainer, gasPrice: toBN("0") }
-      ),
-      "onlyMember"
-    );
-  });
-
-  it("should not be possible to update a DAO configuration if you are a member but not a maintainer", async () => {
-    const maintainer = accounts[5]; // not a member
-
-    // Issue OpenLaw ERC20 Basic Token for tests, only DAO maintainer will hold this token
-    const tokenSupply = toBN(100000);
-    const oltContract = await OLToken.new(tokenSupply);
-
-    // Transfer OLTs to the maintainer account
-    await oltContract.transfer(maintainer, toBN(1));
-    const maintainerBalance = await oltContract.balanceOf.call(maintainer);
-    expect(maintainerBalance.toString()).equal("1");
-
-    const { dao, adapters } = await deployDefaultDao({
-      owner,
-      maintainerTokenAddress: oltContract.address, //only holders of the OLT tokens are considered maintainers
-    });
-    const voting = adapters.voting;
-    const configuration = adapters.configuration;
-    const configKey = sha3(
-      web3.utils.encodePacked(
-        "governance.role.",
-        utils.getAddress(configuration.address)
-      )
-    );
-
-    // Make sure the governance token configuration was created
-    const governanceToken = await dao.getAddressConfiguration(configKey);
-    expect(governanceToken).equal(oltContract.address);
-
-    let key = sha3("key");
-
-    const proposalId = getProposalCounter();
-
-    // The DAO owner submits a new configuration proposal
-    await configuration.submitProposal(
-      dao.address,
-      proposalId,
-      [
-        {
-          key: key,
-          numericValue: 99,
-          addressValue: ZERO_ADDRESS,
-          configType: 0,
-        },
-      ],
-      [],
-      {
-        from: owner, // The DAO Owner is not a maintainer because does not hold any OLT Tokens
-        gasPrice: toBN("0"),
-      }
-    );
-
-    let value = await dao.getConfiguration(key);
-    expect(value.toString()).equal("0");
-
-    // The DAO owner attempts to vote on the new proposal,
-    // but since he is not a maintainer (does not hold OLT tokens) the voting weight is zero
-    // so the vote should not be allowed
-    await expectRevert(
-      voting.submitVote(dao.address, proposalId, 1, {
-        from: owner,
-        gasPrice: toBN("0"),
-      }),
-      "vote not allowed"
     );
   });
 });
